% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoReg.R
\name{LassoReg}
\alias{LassoReg}
\title{LASSO Regression}
\usage{
LassoReg(X, y, lambda = 0.1, soft = TRUE, thresh = 1e-06, max_iter = 100)
}
\arguments{
\item{X}{an n*p design matrix}

\item{y}{an n*1 response vector}

\item{lambda}{a pre-specified tuning parameter}

\item{soft}{logical. If TRUE (the default) use soft-thresholding operator}

\item{thresh}{convergence threshold for coordinate descent}

\item{max_iter}{max number of iterations for coordinate descent}
}
\value{
an p*1 LASSO regression coefficients vector
}
\description{
a function that computes the LASSO regression coefficient estimates given X, y and lambda.
}
\examples{
#example 1 n>p
set.seed(123)
X=matrix(rnorm(100*12),nrow=100,ncol=12)
beta=c(rnorm(2,2,1),rnorm(3,-1,1),rep(0,7))
y=X\%*\%beta+rnorm(100,sd=1)
beta_lasso<-LassoReg(X,y,lambda=0.01,thresh = 1e-7,max_iter = 5000,soft = FALSE)
plot(beta)
points(beta_lasso,col='red')



#example 2 n<p
set.seed(123)
X=matrix(rnorm(10*20),nrow=10,ncol=20)
beta=c(rnorm(2,2,1),rnorm(3,-1,1),rep(0,7),rnorm(4,3,2),rnorm(1,-4,1),rep(0,3))
y=X\%*\%beta+rnorm(10,sd=0.5)
beta_lasso<-LassoReg(X,y,lambda=.1,thresh = 1e-7,max_iter = 5000,soft = FALSE)
plot(beta)
points(beta_lasso,col='red')

}
\author{
Rongqian Zhang
}
