% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RegPath.R
\name{plot.RegPath}
\alias{plot.RegPath}
\title{plot.RegPath}
\usage{
\method{plot}{RegPath}(lambda, beta_hat)
}
\arguments{
\item{lambda}{an m*1 regularization parameter vector}

\item{beta_hat}{an p*m estimated coefficient matrix}
}
\value{
a image showing the coefficient paths given lambda
}
\description{
a function that produces a coefficient profile plot of the coefficient paths given lambda
}
\examples{
data(longley)
lambda=seq(from=0,to=0.1,by=0.01)
beta_hat=matrix(nrow=6,ncol = length(lambda))
for (i in 1:length(lambda))
{
beta_hat[,i]=RidgeReg(scale(longley[,2:7]),longley$GNP.deflator,lambda = lambda[i])
}
plot.RegPath(lambda,beta_hat)

}
\author{
Rongqian Zhang
}
