% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeReg.R
\name{RidgeReg}
\alias{RidgeReg}
\title{Ridge Regression}
\usage{
RidgeReg(X, y, lambda)
}
\arguments{
\item{X}{an n*p design matrix}

\item{y}{an n*1 response vector}

\item{lambda}{a pre-specified tuning parameter}
}
\value{
an p*1 ridge regression coefficients vector
}
\description{
a function that computes the ridge regression coefficient estimates given X, y and lambda.
}
\examples{
#example 1 n>p
set.seed(123)
X=matrix(rnorm(100*80),nrow=100,ncol=80)
beta=rt(80,df=2)
y=X\%*\%beta+rnorm(100,sd=2)
RidgeReg(X,y,1)
plot(beta)
points(RidgeReg(X,y,1),col='red')

#example 2 n<p
set.seed(123)
X=matrix(rnorm(100*80),nrow=80,ncol=100)
beta=rt(100,df=2)
y=X\%*\%beta+rnorm(80,sd=2)
RidgeReg(X,y,1)
plot(beta)
points(RidgeReg(X,y,1),col='red')

}
\author{
Rongqian Zhang
}
